Выносим общую логику обновления
